+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Changelog
 
 **Next release**
 
+* Display whitelist icon when allowed account, applications or contacts
+
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)
 
 **Version 2.0.36 STABLE**